requirements.txt
Flask==1.1.2
Flask-Login==0.5.0
google-cloud-datastore==2.1.0
pycryptodome==3.9.9
#six==1.12.0
python-dateutil==2.8.1
google-cloud-tasks==2.1.0
google-cloud-storage==1.35.0
google-python-cloud-debugger
sortedcontainers==2.3.0
dateparser==1.0.0
pytest==6.2.5
ptest-cov==3.0.0
